Posted changes by flatfoot74
from the basicroleplaying.org forums
"It's updated as best I can find with the 1.02 rtf as the basis and the changes from the 1.02 to 1.03 pdf files. I've also included some of the changes specified in the corrections thread. Some weren't included because I was staying away from rewriting the rules. Just making word changes, spelling and grammar corrections. I even found and corrected some of my own.

There is also one correction to the 1.03 rules from the GM Basic Roleplaying Rules Reference, which is under the same ORC license. The Rules Reference has corrections in the Attack & Defense Matrix that were pointed out in the corrections thread. I've added the corrected wording.

Like the 1.02 rtf, I didn't include most of chapter 1 and didn't include the Appendices.

I also corrected what I saw as a glaring error in 1.03, p 249, Creature Hit Location Table (Continued). Row 100, etc., column 1/5 (or 20%) currently reads 16. If it follows the patter of that column, it should read 17 instead. I corrected that in my updates."

https://basicroleplaying.org/topic/22697-download-the-free-basic-roleplaying-orc-content-document-%E2%80%93-sell-the-games-you-create-royalty-free/page/2/#findComment-401485